Load promises file from assets
